+++ main.py
@@ -183,6 +183,8 @@ while running:
     pc.player_card(screen, players[(turns-1) % len(players)+1])
     pc.win_condition_Card(screen,  players[(turns-1) % len(players)+1])
     
+    r_ui.ruleCard(screen)
+    
     player2.spawn(screen)
 
     dc1.spawnDice(screen)