+++ CHANGELOG.md
@@ -9,6 +9,11 @@ This build is not intended for submission, but rather for the developers to keep
 - Some code has been restructured.
 - That means a fix for the player logic had to be made.
 - The players can move again.
+- The dynamic rules are in development.
+    - The house and hotel prices are now variables which can be changed.
+    - The events "Shop closure", "Housing crisis" and "Housing abundance" have been prepared.
+    - The events "Birthday" and "Come out of jail" have been prepared.
+- The rule card is in preparation.
 
 ## Known issues